+++ Include/MathUtils.h
@@ -4,35 +4,35 @@ namespace photon {
 
     struct Vec2f {
 
-        float x;
-        float y;
+        double x;
+        double y;
 
         Vec2f();
-        Vec2f(float x, float y);
+        Vec2f(double x, double y);
 
     };
 
     struct Vec3f {
 
-        float x;
-        float y;
-        float z;
+        double x;
+        double y;
+        double z;
 
         Vec3f();
-        Vec3f(float x, float y, float z);
+        Vec3f(double x, double y, double z);
 
         Vec3f operator+(const Vec3f& other) const;
         Vec3f operator-() const;
         Vec3f operator-(const Vec3f& other) const;
-        Vec3f operator*(float s) const;
+        Vec3f operator*(double s) const;
         Vec3f operator*(const Vec3f& other) const;
         
         Vec3f& operator+=(const Vec3f& other);
 
-        float length() const;
+        double length() const;
         Vec3f normalize() const;
-        float distance(const Vec3f& other) const;
-        float dot(const Vec3f& other) const;
+        double distance(const Vec3f& other) const;
+        double dot(const Vec3f& other) const;
         Vec3f cross(const Vec3f& other) const;
 
     };
@@ -44,7 +44,7 @@ namespace photon {
 
         Ray(const Vec3f& origin, const Vec3f& direction);
 
-        Vec3f point_at(float t) const;
+        Vec3f point_at(double t) const;
 
     };
