+++ Source/Geometry.cpp
@@ -1,11 +1,12 @@
 #include "Geometry.h"
 
+#include <limits>
+
 namespace photon {
 
     Geometry::Geometry(const Material& material) : material(material) {}
 
-    bool Geometry::is_valid_intersection(float t) const {
-        // TODO: Replace 0.0f with the appropriate distance
+    bool Geometry::is_valid_intersection(double t) const {
         return t > 0.0f;
     }
